requirements.txt
pandas
requests
python-dotenv
google-cloud-storage
google-cloud-bigquery
sqlalchemy
pymysql
cryptography
pyarrow
fastparquet